added mobile environmentCard and isMobilehook

+++ src/app/dashboard/components/EnvironmentTab.tsx
@@ -1,3 +1,4 @@
+"use client"
 import { usePlantMonitor } from '@/context/PlantMonitorContext';
 import { EnvironmentTableCard } from '@/components/Table/TableCard';
 import { useRouter } from 'next/navigation';
@@ -6,6 +7,8 @@ import EmptyState from '@/app/environments/[environmentId]/components/shared/Emp
 import { EnvironmentData } from '@/types/environment';
 import { mapEnvironmentsToTableRows } from '@/components/Table/adapters/environmentTableAdapter';
 import { environmentTableConfig } from '@/config/environmentTableConfig';
+import { useIsMobile } from '@/hooks/useIsMobile';
+import { EnvironmentMobileList } from './EnvironmentMobileCard';
 
 interface EnvironmentTabProps {
     environments: EnvironmentData[];
@@ -14,24 +17,33 @@ interface EnvironmentTabProps {
 
 export default function EnvironmentTab({ environments, onAddNew }: EnvironmentTabProps) {
     const { deletePlants } = usePlantMonitor();
-    console.log('environments', environments);
+    const router = useRouter();
+    const isMobile = useIsMobile();
     const rows = mapEnvironmentsToTableRows(environments);
-    console.log('Mapped Environment Table Rows:', rows);
-    const router = useRouter()
+
+    if (environments.length === 0) {
+        return (
+            <TabContent id="environments">
+                <EmptyState message='Keine Environments vorhanden' />
+            </TabContent>
+        );
+    }
+
     return (
         <TabContent id="environments">
-            {environments.length === 0 ? (
-                <EmptyState message='Keine Environments vorhanden' />
+            {isMobile ? (
+                <EnvironmentMobileList
+                    rows={rows}
+                    onAddNew={onAddNew}
+                />
             ) : (
                 <EnvironmentTableCard
                     data={rows}
                     tableConfig={{
                         ...environmentTableConfig,
                         onDeleteSelected: (keys: string[]) => deletePlants(keys),
-                        onRowClick: (row) => {
-                            router.push(`/environments/${row.key}`);
-                        },
-                        onAddNew
+                        onRowClick: (row) => router.push(`/environments/${row.key}`),
+                        onAddNew,
                     }}
                 />
             )}
